requirements.txt
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.1
DAWG-Python==0.7.2
dj-database-url==0.5.0
Django==2.1.7
django-tastypie==0.14.2
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
docopt==0.6.2
gunicorn==19.9.0
httpie==1.0.2
idna==2.8
nltk==3.4
Pygments==2.3.1
PyJWT==1.7.1
pymorphy2==0.8
pymorphy2-dicts==2.4.393442.3710985
PyQt5==5.11.3
PyQt5-sip==4.19.13
python-dateutil==2.8.0
python-mimeparse==1.6.0
pytz==2018.9
requests==2.21.0
singledispatch==3.4.0.3
six==1.11.0
urllib3==1.24.1
whitenoise==4.1.2
psycopg2==2.7.7
django-cors-headers==2.5.2